sms_backend/requirements/base.txt
# Django Framework
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.2
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.9
django-redis==5.4.0

# Async Tasks
celery==5.3.6
redis==5.0.3
flower==2.0.1

# Validation & Serialization
pydantic==2.7.0
marshmallow==3.21.1

# Documentation
drf-spectacular==0.27.1

# File Handling
Pillow==10.3.0
django-storages==1.14.2
boto3==1.34.0

# Utilities
python-dateutil==2.9.0
python-dotenv==1.0.1
pytz==2024.1
uuid==1.30

# Security
django-ratelimit==4.1.0
django-defender==0.9.7

# Monitoring & Logging
sentry-sdk==1.45.0
django-prometheus==2.3.1

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Health Checks
django-health-check==3.18.1
